% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_to_na.R
\name{mutate_to_na}
\alias{mutate_to_na}
\title{Replace multiple elements to Missing.}
\usage{
mutate_to_na(data, val_to_repl, datatype = "is.double", var = FALSE, variable)
}
\arguments{
\item{data}{- The data set to be operated on}

\item{val_to_repl}{- elements to be replace to NA}

\item{datatype}{- to help replace only a given data type}

\item{var}{- logical condition, if set to TRUE allow you select the variables of interest.}

\item{variable}{- allow you to insert the variable of interest or a vector of variable if var = TRUE.}
}
\description{
Function allows you to replace for multiple elements in the data set to NA
}
