% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_to_na_all.R
\name{replace_to_na_all}
\alias{replace_to_na_all}
\title{Replace all values with NA where a certain condition is met}
\usage{
replace_to_na_all(data, condition)
}
\arguments{
\item{data}{- The data set to be operated on}

\item{condition}{- elements to be replace to NA}
}
\description{
Function allows you to replace for multiple elements in the data set to NA
}
\examples{
df <- tibble::tribble(~x,  ~y,    ~z,
                      1,   "A",   -100,
                      3,   "N/A", -99,
                      NA,  NA,    -98,
                      -99, "E",   -101,
                      -98, "F",   -1)

df
#replace all instances of -99 & N/A with NA
replace_to_na_all(data = df,
                   condition = ~.x == -99)
#replace all instances of -99 with NA
replace_to_na_all(data = df,
                   condition = c(-99,"N/A"))


}
