% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_to_na_at.R
\name{replace_to_na_at}
\alias{replace_to_na_at}
\title{Replace specified variables to NA where a certain condition is met}
\usage{
replace_to_na_at(data, .vars, condition)
}
\arguments{
\item{data}{- The data set to be operated on}

\item{.vars}{- A character string of variables to replace with NA values}

\item{condition}{- A condition required to be TRUE to set NA, all elements to be replace to NA}
}
\description{
Function allows you to replace for multiple elements in the data set to NA
}
\examples{
df <- tibble::tribble(~x,  ~y,    ~z,
                          1,   "A",   -100,
                          3,   "N/A", -99,
                          NA,  NA,    -98,
                          -99, "E",   -101,
                          -98, "F",   -1)

df

#replace all instances of -99 with NA
replace_to_na_at(data = df,
                   .vars = "x",
                   condition = ~.x == -99)
#replace all instances of -99 & N/A with NA
replace_to_na_at(data = df,
                   .vars = c("x","z"),
                   condition = c(-99,"N/A"))


}
