% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_to_na_if.R
\name{replace_to_na_if}
\alias{replace_to_na_if}
\title{Replace values with Missing(NA) based on some condition, for variables that meet some predicate}
\usage{
replace_to_na_if(data, .predicate, condition)
}
\arguments{
\item{data}{- The data set to be operated on}

\item{.predicate}{- A predicate function to be applied to the columns or a logical vector.}

\item{condition}{- A condition required to be TRUE to set NA, all elements to be replace to NA}
}
\description{
Function allows you to replace for multiple elements in the data set to NA if condition required to be TRUE
}
\examples{
df <- tibble::tribble(~x,  ~y,    ~z,
                          1,   "A",   -100,
                          3,   "N/A", -99,
                          NA,  NA,    -98,
                          -99, "E",   -101,
                          -98, "F",   -1)

df

#replace all instances of "N/A","E","F" with NA
replace_to_na_if(data = df,
                   .predicate = is.character,
                   condition = c("N/A","E","F"))
#replace all instances of -99,-98,-99,-1 with NA
replace_to_na_if(data = df,
                   .predicate = is.double,
                 condition = c(-99,-98,-99,-1))


}
