% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified Systematic sampling-Remmap.R
\name{strat_syst_sampling}
\alias{strat_syst_sampling}
\title{Stratified Systematic Sample for sampling AIM 3 subjects from AIM 2 in REMAPP}
\usage{
strat_syst_sampling(data, n, trimester = "Trimester 1", trim_var_exist = FALSE)
}
\arguments{
\item{data}{- The data set to be operated on}

\item{n}{- sample size}

\item{trimester}{- Specify trimester stratum for for sampling i.e. "Trimester 1","Trimester 2" or "Trimester 3".}

\item{trim_var_exist}{- Is TRUE If trimester Variable exist else FALSE for the algorithm to help create one.}
}
\description{
Function allows you to sample AIM3 cohort from AIM 2 using Trimester as strata. for "Trimester 1","Trimester 2" & "Trimester 3"
}
\details{
If you have an existing Trimester stratification variable, you must mutate it to lowercase i.e. trimester.
}
\examples{
#Sampling n= 10 for "Trimester 2"
set.seed(123)
Remapp_data <- data.frame(MOMID = paste0(rep("KEARC000",50),1:50),
US_GA_WKS_AGE_FTS1 = rnorm( mean = 14,sd = 4,50),
US_GA_DAYS_AGE_FTS1 = rep(c(1:5),10),
TYPE_VISIT = c(rep(1,42),rep(6,3),rep(2:5),2),
CON_DSSTDAT = sample(seq(Sys.Date()-60,Sys.Date()-5,1),50,replace = FALSE)
)
df = strat_syst_sampling (data = Remapp_data ,n = 10,trimester = "Trimester 2",trim_var_exist = FALSE)
df

}
